refactor unittest, extract mock componets to separate module
