+++ client/src/app/+dashboard/analog-devices/temperature/all-temperatures/all-temperatures.component.spec.ts
@@ -2,7 +2,7 @@ import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {AllTemperaturesComponent} from './all-temperatures.component';
 import {DeviceType} from '../../../../../../../server/entities/device-type';
-import {Component, Input} from '@angular/core';
+import {TestingMocksModule} from '../../../../testing-mocks/testing-mocks.module';
 
 describe('AllTemperaturesComponent', () => {
   let component: AllTemperaturesComponent;
@@ -10,9 +10,11 @@ describe('AllTemperaturesComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [
+        TestingMocksModule,
+      ],
       declarations: [
         AllTemperaturesComponent,
-        MockAllOfTypeComponent,
       ]
     })
       .compileComponents();
@@ -32,14 +34,5 @@ describe('AllTemperaturesComponent', () => {
     const allOfTypeComponent = fixture.debugElement.children[0];
     expect(allOfTypeComponent.context.deviceType).toBe(DeviceType.TEMPERATURE);
   });
-
 });
-
-@Component({
-  selector: 'app-all-of-type',
-  template: '<div></div>',
-})
-class MockAllOfTypeComponent {
-  @Input() deviceType: string;
-}
